+++ CHANGELOG.md
@@ -23,6 +23,7 @@ the detailed section referring to by linking pull requests or issues.
 - Copyable contact email and subject fields on asset and data offer detail
   dialogs
 - Assets Page search input field is now case insensitive
+- Performance improvement when fetching a single contract agreement
 - Temporary reimplement the Edit Asset as a Dialog
 
 ## [v4.1.0] - 2024-07-24